refactor added company dto with basic info
